split parser; normalize domain
