+++ adstxt.go
@@ -2,12 +2,8 @@
 package adstxt
 
 import (
-	"bufio"
-	"fmt"
 	"io"
 	"net/http"
-	"regexp"
-	"strings"
 )
 
 // Record is ads.txt data field defined in iab.
@@ -44,75 +40,16 @@ func Get(rawurl string) ([]Record, error) {
 	return Parse(resp.Body)
 }
 
-func parseAccountType(s string) AccountType {
-	switch strings.ToUpper(s) {
-	case "DIRECT":
-		return AccountDirect
-	case "RESELLER":
-		return AccountReseller
-	default:
-		// NOTE or should be error ?
-		return AccountOther
-	}
-}
-
-var leadingBlankRe = regexp.MustCompile(`\A[\s\t]+`)
-var trailingBlankRe = regexp.MustCompile(`[\s\t]+\z`)
-
-func normalize(s string) string {
-	// sanitize blank characters
-	s = leadingBlankRe.ReplaceAllString(s, "")
-	s = trailingBlankRe.ReplaceAllString(s, "")
-	return s
-}
-
-func parseRow(row string) (*Record, error) {
-	// dropping extension field
-	if idx := strings.Index(row, ";"); idx != -1 {
-		row = row[0:idx]
-	}
-
-	fields := strings.Split(row, ",")
-
-	// if the first field contains "=", then the row is for key-value definitions
-	if strings.Index(fields[0], "=") != -1 {
-		return nil, nil
-	}
-
-	if l := len(fields); l != 3 && l != 4 {
-		return nil, fmt.Errorf("ads.txt has fields length is incorrect.: %s", row)
-	}
-
-	// otherwise the row is valid
-	var r Record
-	r.ExchangeDomain = normalize(fields[0])
-	r.PublisherAccountID = normalize(fields[1])
-	r.AccountType = parseAccountType(normalize(fields[2]))
-	// AuthorityID is optional
-	if len(fields) >= 4 {
-		r.AuthorityID = normalize(fields[3])
-	}
-	return &r, nil
-}
-
 func Parse(in io.Reader) ([]Record, error) {
 	records := make([]Record, 0)
-	scanner := bufio.NewScanner(in)
-	for scanner.Scan() {
-		text := scanner.Text()
+	p := NewParser(in)
 
-		// blank line
-		if len(text) == 0 {
-			continue
+LOOP:
+	for {
+		r, err := p.Parse()
+		if err == io.EOF {
+			break LOOP
 		}
-
-		// comment out
-		if []rune(text)[0] == '#' {
-			continue
-		}
-
-		// otherwise try parsing
-		r, err := parseRow(scanner.Text())
 		if err != nil {
 			return nil, err
 		}
@@ -120,8 +57,6 @@ func Parse(in io.Reader) ([]Record, error) {
 			records = append(records, *r)
 		}
 	}
-	if err := scanner.Err(); err != nil {
-		return nil, err
-	}
+
 	return records, nil
 }
